requirements.txt
alembic==1.7.4
amqp==5.0.6
aniso8601==9.0.1
appnope==0.1.2
autopep8==1.5.7
aws==0.2.5
awscli==1.20.60
backcall==0.2.0
bcrypt==3.2.0
billiard==3.6.4.0
boto==2.49.0
boto3==1.18.59
botocore==1.21.60
cachetools==4.2.4
celery==5.1.2
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
cryptography==35.0.0
debugpy==1.5.0
decorator==5.1.0
docutils==0.15.2
entrypoints==0.3
fabric==2.6.0
Flask==2.0.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
google-api-core==2.1.1
google-api-python-client==2.26.1
google-auth==2.3.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
greenlet==1.1.2
httplib2==0.20.1
idna==3.3
invoke==1.6.0
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.2
jmespath==0.10.0
jupyter-client==7.0.6
jupyter-core==4.8.1
kombu==5.1.0
Mako==1.1.5
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.1
numpy==1.21.2
oauthlib==3.1.1
pandas==1.3.3
paramiko==2.8.0
parso==0.8.2
pathlib2==2.3.6
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
prettytable==2.2.1
prompt-toolkit==3.0.20
protobuf==3.18.1
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.10.0
PyJWT==2.2.0
PyNaCl==1.4.0
pyodbc==4.0.32
pyparsing==2.4.7
pypyodbc==1.3.2
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.3
PyYAML==5.4.1
pyzmq==22.3.0
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.5.0
six==1.16.0
SQLAlchemy==1.4.25
toml==0.10.2
tornado==6.1
traitlets==5.1.0
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.2
